pseudo wav pipeline as shell script
